chore(fileStruct) Splitted SampleProviders in separate files.
